REFACTORING: More understandable components folders structure.
