+++ src/components/InputWithSuggestions.jsx
